add mybatis-plus

+++ src/main/java/com/example/rbacdemo/service/UserService.java
@@ -1,9 +1,6 @@
 package com.example.rbacdemo.service;
 
-import java.util.List;
-
 import com.baomidou.mybatisplus.core.metadata.IPage;
-import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.example.rbacdemo.common.PageQuery;
 import com.example.rbacdemo.model.SysUser;
 
@@ -12,7 +9,7 @@ public interface UserService {
 
   public IPage<SysUser> getUserList(PageQuery pageQuery);
 
-  public Long getTotal();
+  public Integer getTotal();
 
   public void update(SysUser user);
 }
